Vibrator: add a mutex in vibration play to avoid race condition
The composition effects with primitives are played in child thread and
it calls play() function which would be also called by the main thread
if it wants to stop the play at the same time, this introduces a race so
add a mutex lock in play() to avoid it.

Meanwhile, clear the event pipe before triggering a new play regardless
of whether there was an unfinished composition play, this helps to clear
the event if it was cached in the pipe unexpectedly in other corner
cases, such as, if the event was sent while the composePlayThread just
waited it timeout, the event will be kept in the pipe and it needs to be
cleared before starting a new play.

CRs-Fixed: 3426813
Change-Id: Id1e2f004626f0aee1531fa43442e0e3e0d2d25e7